Inventory Updates

+++ backend/src/main/groovy/com/backend/gbp/domain/inventory/BeginningBalance.groovy
@@ -47,7 +47,7 @@ class BeginningBalance extends AbstractAuditingEntity {
 
 	@GraphQLQuery
 	@Column(name = "quantity")
-	Integer quantity
+	BigDecimal quantity
 
 	@GraphQLQuery
 	@Column(name = "unit_cost")
